Consolidate URL_LIMITS, url_limit() into utils, conftest
