+++ tiled/_tests/conftest.py
@@ -10,7 +10,9 @@ from sqlalchemy.ext.asyncio import create_async_engine
 
 from .. import profiles
 from ..catalog import from_uri, in_memory
+from ..client.base import BaseClient
 from ..server.settings import get_settings
+from .utils import URL_LIMITS
 from .utils import enter_password as utils_enter_password
 from .utils import temp_postgres
 
@@ -217,3 +219,18 @@ def set_and_deplete(request: pytest.FixtureRequest):
     values = set(request.param)
     yield values
     assert not values
+
+
+@pytest.fixture
+def url_limit(request: pytest.FixtureRequest):
+    """Adjust the URL length limit for the client's GET requests.
+
+    Data will be fetched by GET requests when the URL_CHARACTER_LIMIT is long,
+    and by equivalent POST requests when the URL_CHARACTER_LIMIT is short.
+    """
+    URL_CHARACTER_LIMIT = int(request.param)
+    # Temporarily adjust the URL length limit to change client behavior.
+    BaseClient.URL_CHARACTER_LIMIT = URL_CHARACTER_LIMIT
+    yield
+    # Then restore the original value.
+    BaseClient.URL_CHARACTER_LIMIT = int(URL_LIMITS.ORIGINAL)
